rename library module
