+++ angular.json
@@ -3,7 +3,7 @@
   "version": 1,
   "newProjectRoot": "projects",
   "projects": {
-    "next-assay-pre-approved-study-details-lib-app": {
+    "next-assay-pre-approved-study-details-app": {
       "root": "",
       "sourceRoot": "src",
       "projectType": "application",
@@ -17,7 +17,7 @@
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
-            "outputPath": "dist/next-assay-pre-approved-study-details-lib-app",
+            "outputPath": "dist/next-assay-pre-approved-study-details-app",
             "index": "src/index.html",
             "main": "src/main.ts",
             "polyfills": "src/polyfills.ts",
@@ -61,18 +61,18 @@
         "serve": {
           "builder": "@angular-devkit/build-angular:dev-server",
           "options": {
-            "browserTarget": "next-assay-pre-approved-study-details-lib-app:build"
+            "browserTarget": "next-assay-pre-approved-study-details-app:build"
           },
           "configurations": {
             "production": {
-              "browserTarget": "next-assay-pre-approved-study-details-lib-app:build:production"
+              "browserTarget": "next-assay-pre-approved-study-details-app:build:production"
             }
           }
         },
         "extract-i18n": {
           "builder": "@angular-devkit/build-angular:extract-i18n",
           "options": {
-            "browserTarget": "next-assay-pre-approved-study-details-lib-app:build"
+            "browserTarget": "next-assay-pre-approved-study-details-app:build"
           }
         },
         "test": {
@@ -106,7 +106,7 @@
         }
       }
     },
-    "next-assay-pre-approved-study-details-lib-app-e2e": {
+    "next-assay-pre-approved-study-details-app-e2e": {
       "root": "e2e/",
       "projectType": "application",
       "prefix": "",
@@ -115,11 +115,11 @@
           "builder": "@angular-devkit/build-angular:protractor",
           "options": {
             "protractorConfig": "e2e/protractor.conf.js",
-            "devServerTarget": "next-assay-pre-approved-study-details-lib-app:serve"
+            "devServerTarget": "next-assay-pre-approved-study-details-app:serve"
           },
           "configurations": {
             "production": {
-              "devServerTarget": "next-assay-pre-approved-study-details-lib-app:serve:production"
+              "devServerTarget": "next-assay-pre-approved-study-details-app:serve:production"
             }
           }
         },
@@ -134,33 +134,33 @@
         }
       }
     },
-    "next-assay-pre-approved-study-details-lib": {
-      "root": "projects/next-assay-pre-approved-study-details-lib",
-      "sourceRoot": "projects/next-assay-pre-approved-study-details-lib/src",
+    "next-assay-pre-approved-study-details": {
+      "root": "projects/next-assay-pre-approved-study-details",
+      "sourceRoot": "projects/next-assay-pre-approved-study-details/src",
       "projectType": "library",
       "prefix": "next",
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-ng-packagr:build",
           "options": {
-            "tsConfig": "projects/next-assay-pre-approved-study-details-lib/tsconfig.lib.json",
-            "project": "projects/next-assay-pre-approved-study-details-lib/ng-package.json"
+            "tsConfig": "projects/next-assay-pre-approved-study-details/tsconfig.lib.json",
+            "project": "projects/next-assay-pre-approved-study-details/ng-package.json"
           }
         },
         "test": {
           "builder": "@angular-devkit/build-angular:karma",
           "options": {
-            "main": "projects/next-assay-pre-approved-study-details-lib/src/test.ts",
-            "tsConfig": "projects/next-assay-pre-approved-study-details-lib/tsconfig.spec.json",
-            "karmaConfig": "projects/next-assay-pre-approved-study-details-lib/karma.conf.js"
+            "main": "projects/next-assay-pre-approved-study-details/src/test.ts",
+            "tsConfig": "projects/next-assay-pre-approved-study-details/tsconfig.spec.json",
+            "karmaConfig": "projects/next-assay-pre-approved-study-details/karma.conf.js"
           }
         },
         "lint": {
           "builder": "@angular-devkit/build-angular:tslint",
           "options": {
             "tsConfig": [
-              "projects/next-assay-pre-approved-study-details-lib/tsconfig.lib.json",
-              "projects/next-assay-pre-approved-study-details-lib/tsconfig.spec.json"
+              "projects/next-assay-pre-approved-study-details/tsconfig.lib.json",
+              "projects/next-assay-pre-approved-study-details/tsconfig.spec.json"
             ],
             "exclude": [
               "**/node_modules/**"
@@ -170,5 +170,5 @@
       }
     }
   },
-  "defaultProject": "next-assay-pre-approved-study-details-lib-app"
+  "defaultProject": "next-assay-pre-approved-study-details-app"
 }
