+++ e2e/src/app.e2e-spec.ts
@@ -9,6 +9,6 @@ describe('workspace-project App', () => {
 
   it('should display welcome message', () => {
     page.navigateTo();
-    expect(page.getTitleText()).toEqual('Welcome to next-assay-pre-approved-study-details-lib-app!');
+    expect(page.getTitleText()).toEqual('Welcome to next-assay-pre-approved-study-details-app!');
   });
 });
