+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "next-assay-pre-approved-study-details-lib-app",
+  "name": "next-assay-pre-approved-study-details-app",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
