--- projects/next-assay-pre-approved-study-details-lib/src/public_api.ts
@@ -1,6 +1,0 @@
-/*
- * Public API Surface of next-assay-pre-approved-study-details-lib
- */
-
-export * from './lib/next-assay-pre-approved-study-details/next-assay-pre-approved-study-details.component';
-export * from './lib/next-assay-pre-approved-study-details-lib.module';
