+++ projects/next-assay-pre-approved-study-details/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "next-assay-pre-approved-study-details-lib",
+  "name": "next-assay-pre-approved-study-details",
   "version": "0.0.1",
   "peerDependencies": {
     "@angular/common": "^7.1.0",
