+++ projects/next-assay-pre-approved-study-details/src/public_api.ts
@@ -1,0 +1,6 @@
+/*
+ * Public API Surface of next-assay-pre-approved-study-details
+ */
+
+export * from './lib/next-assay-pre-approved-study-details/next-assay-pre-approved-study-details.component';
+export * from './lib/next-assay-pre-approved-study-details.module';
