+++ tsconfig.json
@@ -19,11 +19,11 @@
       "dom"
     ],
     "paths": {
-      "next-assay-pre-approved-study-details-lib": [
-        "dist/next-assay-pre-approved-study-details-lib"
+      "next-assay-pre-approved-study-details": [
+        "dist/next-assay-pre-approved-study-details"
       ],
-      "next-assay-pre-approved-study-details-lib/*": [
-        "dist/next-assay-pre-approved-study-details-lib/*"
+      "next-assay-pre-approved-study-details/*": [
+        "dist/next-assay-pre-approved-study-details/*"
       ]
     }
   }
